pyproject.toml
[project]
name = "overseerr-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.6.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.3",
]

[project.scripts]
overseerr-mcp = "overseerr_mcp:main"
